Only patch in unittests
